+++ src/IECoreArnold/Renderer.cpp
@@ -1255,20 +1255,6 @@ IECore::InternedString g_lightFilterPrefix( "ai:lightFilter:" );
 
 IECore::InternedString g_filteredLights( "filteredLights" );
 
-IECoreScene::ConstShaderNetworkPtr g_facingRatio = []() {
-
-	IECoreScene::ShaderNetworkPtr result = new IECoreScene::ShaderNetwork;
-
-	const IECore::InternedString utilityHandle = result->addShader(
-		"utility", new IECoreScene::Shader( "utility" )
-	);
-
-	result->setOutput( { utilityHandle, "out" } );
-
-	return result;
-
-} ();
-
 const std::vector<IECore::InternedString> g_surfaceShaderAttributeNames = {
 	"ai:surface",
 	"osl:surface",
@@ -1337,11 +1323,6 @@ class ArnoldAttributes : public IECoreScenePreview::Renderer::AttributesInterfac
 			updateShadingFlag( g_arnoldMatteAttributeName, Matte, attributes );
 
 			m_surfaceShader = shaderCache->get( g_surfaceShaderAttributeNames, attributes );
-			if( !m_surfaceShader )
-			{
-				m_surfaceShader = shaderCache->get( g_facingRatio.get(), "", nullptr );
-			}
-
 			m_volumeShader = shaderCache->get( g_volumeShaderAttributeNames, attributes );
 			m_filterMap = shaderCache->get( g_filterMapAttributeNames, attributes );
 			m_uvRemap = shaderCache->get( g_uvRemapAttributeNames, attributes );
@@ -1663,6 +1644,16 @@ class ArnoldAttributes : public IECoreScenePreview::Renderer::AttributesInterfac
 				else
 				{
 					AiNodeResetParameter( node, g_shaderArnoldString );
+					// `AiNodeResetParameter()` should be sufficient, but recent
+					// Arnold versions seem to have a bug where it doesn't take
+					// effect even though it has the applied the correct value.
+					// Kick it by applying again via `AiNodeSetPtr()`. But don't
+					// do that for procedurals because then the internal shaders
+					// get clobbered by the external one.
+					if( !AiNodeIs( node, g_proceduralArnoldString ) )
+					{
+						AiNodeSetPtr( node, g_shaderArnoldString, AiNodeGetPtr( node, g_shaderArnoldString ) );
+					}
 				}
 
 				if( m_traceSets && m_traceSets->readable().size() )
@@ -1767,7 +1758,7 @@ class ArnoldAttributes : public IECoreScenePreview::Renderer::AttributesInterfac
 
 			// Otherwise use the surface shader. We use this even for volume geometry,
 			// because Gaffer historically assigned volume shaders as `ai:surface`.
-			return m_surfaceShader->root();
+			return m_surfaceShader ? m_surfaceShader->root() : nullptr;
 		}
 
 		const IECoreScene::ShaderNetwork *lightShader() const